Save callback references as GC roots
This wraps the callback in an extra struct which allows retrieval of the
address to be able to be registered as GC root.

What it also requires is to free it afterwards, so these GC roots don't
stick around forever, so some code was added to discard them in the
relevant places.

Closes #5.